requirements.txt
flask>=3.1.0
python-dotenv>=1.0.0
openai>=1.3.0
psycopg2-binary>=2.9.10
pyyaml>=6.0.2
Werkzeug>=3.0.1 